move files into src folder
