+++ src/components/CenteredHero.js
