+++ src/components/images/hero-image.png
