+++ src/pages/privacy.js
