+++ src/pages/terms.js
